+++ Control/Monad/Trans/Random.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE Trustworthy #-}
 
 {- |
-Module       :  Control.Monad.Random
+Module       :  Control.Monad.Trans.Random
 Copyright    :  (c) Cale Gibbard 2006-2007,
                 (c) Russell O'Connor, Dan Doel and Remi Turk 2006,
                 (c) Eric Kidd 2007
@@ -11,11 +11,14 @@ Maintainer   :  byorgey@gmail.com
 Stability    :  experimental
 Portability  :  non-portable (multi-param classes, functional dependencies, undecidable instances)
 
-Declaration of MonadRandom and MonadSplit classes.
+Random monads, passing a random number generator through a computation.
+
+This version is lazy; for a strict version, see
+"Control.Monad.Trans.Random.Strict", which has the same interface.
 -}
 
-module Control.Monad.Random
-    ( module Control.Monad.Random.Lazy,
+module Control.Monad.Trans.Random
+    ( module Control.Monad.Trans.Random.Lazy,
     ) where
 
-import Control.Monad.Random.Lazy
+import Control.Monad.Trans.Random.Lazy